Use uint32_t for secondary points.

+++ include/hyperdex/search.h
@@ -71,8 +71,8 @@ class search
             return m_values.size();
         }
 
-        uint64_t secondary_point() const;
-        uint64_t secondary_mask() const;
+        uint32_t secondary_point() const;
+        uint32_t secondary_mask() const;
 
     private:
         friend e::packer& operator << (e::packer& lhs, const search& rhs);
